Outsource enum tests into an own dedicated test class
